requirements.txt
numpy
pandas
setuptools
google-cloud-storage
pyyaml
scikit_learn
tensorflow
comet-ml
dvc